Add source release timestamps to nodes and edges, pt1
Adds released and expired dates at the data source to the nodes and edges
to allow for matching release times for other data sets (say RefSeq).
Basically analogous to the expired and created fields, but the expired
and created fields refer to when the data was loaded into
our system, not when the data was released at the source.

Change will be in several parts; this is the first.